# Conflicts:
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/NakshaCollection.kt
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/NakshaSession.kt
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/Static.kt
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/write/NakshaBulkLoaderPlan.kt
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/write/NakshaRequestOp.kt
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/write/RowUpdater.kt
#	here-naksha-lib-plv8/src/commonMain/kotlin/naksha/plv8/write/SingleCollectionWriter.kt
#	here-naksha-lib-plv8/src/jvmMain/kotlin/JvmPlv8ResultSet.kt
#	here-naksha-lib-plv8/src/jvmMain/kotlin/JvmPlv8Sql.kt
#	here-naksha-lib-plv8/src/jvmTest/kotlin/naksha/plv8/WriteFeaturesTest.kt

+++ here-naksha-lib-model/src/commonMain/kotlin/naksha/model/response/Metadata.kt
@@ -16,7 +16,7 @@ data class Metadata(
     var txnNext: Int64?,
     var txn: Int64,
     var ptxn: Int64?,
-    var fnva1: Int64?,
+    var fnva1: Int?,
     var uid: Int?,
     var puid: Int?,
     var version: Int?,
